+++ IT326CommunityCalendar/Utils/User_Details.py
@@ -1,4 +1,3 @@
-from Location import Location
 
 class User_Details:
     def __init__ (self,hobbies,age,intrests,gender):
@@ -6,17 +5,7 @@ class User_Details:
         self.age=age
         self.intrests=intrests
         self.gender=gender
-        self.loc=Location("","","",0)
-    
-    def set_location(self,street,city,state,zipcode):
-        self.loc.set_street(street)
-        self.loc.set_city(city)
-        self.loc.set_state(state)
-        self.loc.set_zip(zipcode)
-
-    def get_location(self):
-        location=self.loc.get_street()+" "+ self.loc.get_city()+" "+self.loc.get_state()+" "+self.loc.get_zipcode()
-        return location
+        
     
     def get_hobbies(self):
         return self.hobbies
